+++ mixtape/players.py
@@ -1,8 +1,14 @@
-import asyncio
-import logging
-from typing import Any, Optional, TypeVar, Tuple, List, Callable, MutableMapping
+from typing import (
+    Any,
+    Optional,
+    Type,
+    TypeVar,
+    Tuple,
+    List,
+)
 
 import attr
+import pluggy
 
 import gi
 
@@ -10,200 +16,108 @@ gi.require_version("Gst", "1.0")
 from gi.repository import Gst  # type: ignore
 
 from .base import BasePlayer
-from .exceptions import PlayerPipelineError, PlayerNotConfigured
 from .events import PlayerEvents
+from .plugins import MixtapePlugin, Logger, MessageHandler
+from .spec import PlayerSpec
 
-logger = logging.getLogger(__name__)
 
-AsyncPlayerType = TypeVar("AsyncPlayerType", bound="AsyncPlayer")
+PlayerT = TypeVar("PlayerT", bound="Player")
+MixtapePluginT = Type[MixtapePlugin]
 
 
 @attr.s
-class AsyncPlayer(BasePlayer):
+class Player(BasePlayer):
     """
-    A asyncio compatible player.
+    An asyncio compatible player.
     Interfaces with the `Gst.Bus` with an asyncio file descriptor,
     which is used to set `asyncio.Event` when received for the bus,
     allowing for asyncio compatible methods.
     """
 
-    futures: List[Any] = attr.ib(init=False, default=attr.Factory(list))
+    PLUGIN_SPEC: Any = PlayerSpec
+    DEFAULT_PLUGINS: List[MixtapePluginT] = [Logger]
+    REQUIRED_PLUGINS: List[MixtapePluginT] = [MessageHandler]
+
+    plugins: Type[pluggy.PluginManager] = attr.ib(init=False, repr=False)
     events: PlayerEvents = attr.ib(init=False, default=attr.Factory(PlayerEvents))
-    loop: Optional[asyncio.AbstractEventLoop] = attr.ib(init=False, default=None, repr=False)
-    pollfd: Any = attr.ib(init=False, default=None, repr=False)
-    handlers: MutableMapping[Gst.MessageType, Callable[[Gst.Bus, Gst.Message], None]] = attr.ib(
-        init=False, repr=False
-    )
 
-    # -- defaults -- #
+    @plugins.default
+    def _plugins(self) -> Type[pluggy.PluginManager]:
+        pm = pluggy.PluginManager("mixtape")
+        pm.add_hookspecs(self.PLUGIN_SPEC)
+        return pm
 
-    @handlers.default
-    def _handlers(self) -> MutableMapping[Gst.MessageType, Callable[[Gst.Bus, Gst.Message], None]]:
-        return {
-            Gst.MessageType.ERROR: self.on_error,
-            Gst.MessageType.EOS: self.on_eos,
-            Gst.MessageType.STATE_CHANGED: self.on_state_changed,
-            Gst.MessageType.ASYNC_DONE: self.on_async_done,
-        }
+    @property
+    def hook(self) -> Any:
+        """Convenience shortcut for pm hook"""
+        return self.plugins.hook
 
-    # -- overriding base state shortcuts and pipeline methods with asyncio -- #
-
-    async def ready(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:  # type: ignore
+    async def ready(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:
         """Async override of base.ready"""
-        ret = super().ready()
+        ret = super()._ready()
         await self.events.wait_for_state(Gst.State.READY)
+        self.hook.on_ready()
         return ret
 
-    async def play(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:  # type: ignore
+    async def play(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:
         """Async override of base.play"""
-        ret = super().play()
+        ret = super()._play()
         await self.events.wait_for_state(Gst.State.PLAYING)
+        self.hook.on_play()
         return ret
 
-    async def pause(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:  # type: ignore
+    async def pause(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:
         """Async override of base.pause"""
-        ret = super().pause()
+        ret = super()._pause()
         await self.events.wait_for_state(Gst.State.PAUSED)
+        self.hook.on_pause()
         return ret
 
-    # fmt: off
-    async def stop(self, send_eos: bool = True, teardown: bool = False) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:  # type: ignore
+    async def stop(self) -> Tuple[Gst.StateChangeReturn, Gst.State, Gst.State]:
         """Async override of base.stop"""
-    # fmt: on
-        if send_eos:
-            await self.send_eos()
         ret = self.set_state(Gst.State.NULL)
-
-        if teardown:
-            self.teardown()
-
+        self.hook.on_stop()
         return ret
 
-    async def send_eos(self) -> bool:  # type: ignore
+    async def send_eos(self) -> bool:
+        """Send eos to pipeline and await event"""
         ret = self.pipeline.send_event(Gst.Event.new_eos())
         await self.events.eos.wait()
+        self.hook.on_eos()
         return ret
-
-    # -- bus message handling -- #
-
-    def handle(self) -> None:
-        """
-        Asyncio reader callback, called when a message is available on
-        the bus.
-        """
-        msg = self.bus.pop()
-        if msg:
-            handler = self.handlers.get(msg.type, self.on_unhandled_msg)
-            handler(self.bus, msg)
-
-    def on_state_changed(self, bus: Gst.Bus, message: Gst.Message) -> None:  # pylint: disable=unused-argument
-        """
-        Handler for `state_changed` messages
-        By default will only log to `debug`
-        """
-        old, new, _ = message.parse_state_changed()
-
-        if message.src != self.pipeline:
-            return
-        logger.debug(
-            "State changed from %s to %s",
-            Gst.Element.state_get_name(old),
-            Gst.Element.state_get_name(new),
-        )
-
-        self.events.pick_state(new)
-
-    def on_error(self, bus: Gst.Bus, message: Gst.Message) -> None:
-        """
-        Handler for `error` messages
-        By default it will parse the error message,
-        log to `error` and append to `self.errors`
-        """
-        err, debug = message.parse_error()
-        self.events.error.set()
-        logger.error("Error received from element %s:%s on %s", message.src.get_name(), err.message, bus)
-        if debug is not None:
-            logger.error("Debugging information: %s", debug)
-        raise PlayerPipelineError(err)
-
-    def on_eos(self, bus: Gst.Bus, message: Gst.Message) -> None:
-        """
-        Handler for eos messages
-        By default it sets the eos event
-        """
-        self.events.eos.set()
-        logger.info("EOS message: %s received from pipeline on %s", message, bus)
-
-    def on_async_done(self, bus: Gst.Bus, message: Gst.Message) -> None:
-        """
-        Handler for `async_done` messages
-        By default, it will pop any futures available in `self.futures`
-        and call their result.
-        """
-        msg = message.parse_async_done()
-        logger.debug("Unhandled ASYNC_DONE message: %s on %s", msg, bus)
-
-    def on_unhandled_msg(self, bus: Gst.Bus, message: Gst.Message) -> None:
-        """
-        Handler for all other messages.
-        By default will just log with `debug`
-        """
-        logger.debug("Unhandled msg: %s on %s", message.type, bus)
 
     # -- setup and teaddown -- #
 
     def setup(self) -> None:
         """Setup needs a running asyncio loop"""
+        self.hook.setup()
         super().setup()
-        self.loop = asyncio.get_running_loop()
-        self.pollfd = self.bus.get_pollfd()
-        self.loop.add_reader(self.pollfd.fd, self.handle)
         self.events.setup.set()
-        logger.debug("Setup complete")
 
     def teardown(self) -> None:
         """Cleanup player references to loop and gst resources"""
+        self.hook.teardown()
         super().teardown()
-        if self.loop:
-            self.loop.remove_reader(self.pollfd.fd)
-        self.pollfd = None
-        self.loop = None
         self.events.teardown.set()
-        logger.debug("Teardown complete")
 
-    # -- sync helpers -- #
+    @classmethod
+    async def create(
+        cls: Type[PlayerT], pipeline: Gst.Pipeline, plugins: Optional[List[MixtapePluginT]] = None
+    ) -> PlayerT:
+        """Player factory from a given pipeline that calls setup by default"""
+        player = cls(pipeline)
+        if plugins is None:
+            plugins = cls.REQUIRED_PLUGINS + cls.DEFAULT_PLUGINS
+        for plugin in plugins:
+            player.plugins.register(plugin(player))
+        player.setup()
+        return player
 
-    def call_play(self) -> Any:
-        """Convencience method to call play from another thread"""
-        if self.loop:
-            return asyncio.run_coroutine_threadsafe(self.play(), loop=self.loop)
-        else:
-            raise PlayerNotConfigured("Player setup without running loop not allowed")
-
-    def call_stop(self) -> Any:
-        """Convencience method to call stop from another thread"""
-        if self.loop:
-            return asyncio.run_coroutine_threadsafe(self.stop(), loop=self.loop)
-        else:
-            raise PlayerNotConfigured("Player setup without running loop not allowed")
-
-    def call_pause(self) -> Any:
-        """Convencience method to call pause from another thread"""
-        if self.loop:
-            return asyncio.run_coroutine_threadsafe(self.pause(), loop=self.loop)
-        else:
-            raise PlayerNotConfigured("Player setup without running loop not allowed")
-
-    # -- util -- #
-
-    def create_async_future(self):  # type: ignore
-        """
-        Creates a future and appends it to `futures`.
-        Will `result` with a async_done message.
-        """
-        if self.loop:
-            ft = self.loop.create_future()
-        else:
-            raise PlayerNotConfigured("Player setup without running loop not allowed")
-        self.futures.append(ft)
-        return ft
+    @classmethod
+    async def from_description(
+        cls: Type[PlayerT], description: str, plugins: Optional[List[MixtapePluginT]] = None
+    ) -> PlayerT:
+        """Player factory from a pipeline description"""
+        pipeline = Gst.parse_launch(description)
+        assert isinstance(pipeline, Gst.Pipeline)
+        return await cls.create(pipeline=pipeline, plugins=plugins)