+++ tests/conftest.py
@@ -4,26 +4,20 @@ import colorlog
 import pytest
 
 
-logger = logging.getLogger(__name__)
-handler = colorlog.StreamHandler()
-handler.setFormatter(
-    colorlog.ColoredFormatter(
-        "(%(asctime)s) [%(log_color)s%(levelname)s] | %(name)s | %(message)s [%(threadName)-10s]"
-    )
-)
-
-# get root logger
-logger = logging.getLogger()
-logger.handlers = []
-logger.addHandler(handler)
-logger.setLevel(logging.DEBUG)
-
-
-# flake8 plugin is way too verbose
 def pytest_configure(config):
-    logging.getLogger("flake8").setLevel(logging.WARN)
-    logging.getLogger("bandit").setLevel(logging.WARN)
-    logging.getLogger("blib2to3").setLevel(logging.WARN)
+    handler = colorlog.StreamHandler()
+    handler.setFormatter(
+        colorlog.ColoredFormatter(
+            "(%(asctime)s) [%(log_color)s%(levelname)s] | %(name)s | %(message)s [%(threadName)-10s]"
+        )
+    )
+
+    # get root logger
+    logger = logging.getLogger()
+    logger.handlers = []
+    logger.addHandler(handler)
+    logger.setLevel(logging.WARN)
+    logging.getLogger("mixtape").setLevel(logging.DEBUG)
 
 
 @pytest.fixture
@@ -39,6 +33,6 @@ def Gst():  # noqa
 
 @pytest.fixture
 def player(Gst):
-    from mixtape.players import AsyncPlayer
+    from mixtape import Player
 
-    return AsyncPlayer
+    return Player